Move starting class to package mops
