+++ gruppenbildung/src/main/java/mops/GruppenbildungApplication.java
@@ -1,4 +1,4 @@
-package mops.gruppen1;
+package mops;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class GruppenbildungApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(GruppenbildungApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(GruppenbildungApplication.class, args);
+    }
 
 }
