+++ gruppenbildung/src/test/java/mops/GruppenbildungApplicationTests.java
@@ -1,4 +1,4 @@
-package mops.gruppen1;
+package mops;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class GruppenbildungApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
